django/requirements_legacy_python3.9.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via vdirsyncer
aiosignal==1.3.2
    # via aiohttp
aiostream==0.4.5
    # via vdirsyncer
amqp==5.3.1
    # via kombu
anyascii==0.3.2
    # via wagtail
appy==1.0.19
    # via -r requirements.in
arrow==1.3.0
    # via djangosaml2idp
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
astroid==3.3.11
    # via pylint
async-timeout==5.0.1
    # via redis
atomicwrites==1.4.1
    # via vdirsyncer
attrs==25.3.0
    # via aiohttp
beautifulsoup4==4.11.2
    # via wagtail
billiard==4.2.1
    # via celery
btrees==6.1
    # via zodb
caldav==1.4.0
    # via -r requirements.in
celery==5.5.2
    # via -r requirements.in
certifi==2025.4.26
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   persistent
chardet==5.2.0
    # via reportlab
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-log
    #   click-plugins
    #   click-repl
    #   piecash
    #   vdirsyncer
    #   x-wr-timezone
click-didyoumean==0.3.1
    # via celery
click-log==0.4.0
    # via vdirsyncer
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
contourpy==1.3.0
    # via matplotlib
coverage==7.8.0
    # via -r requirements.in
createsend==9.1.3
    # via -r requirements.in
cryptography==43.0.3
    # via
    #   jwcrypto
    #   pyopenssl
    #   pysaml2
cssselect2==0.8.0
    # via svglib
cycler==0.12.1
    # via matplotlib
datetime==5.5
    # via appy
defusedxml==0.7.1
    # via
    #   pysaml2
    #   willow
dill==0.4.0
    # via pylint
dj-rest-auth==7.0.1
    # via -r requirements.in
django==4.2.21
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-cors-headers
    #   django-datetime-widget
    #   django-filer
    #   django-filter
    #   django-modelcluster
    #   django-oauth-toolkit
    #   django-permissionedforms
    #   django-polymorphic
    #   django-tables2
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   djangosaml2idp
    #   easy-thumbnails
    #   wagtail
django-cors-headers==4.7.0
    # via -r requirements.in
django-datetime-widget==0.9.3
    # via -r requirements.in
django-email-obfuscator==0.1.5
    # via -r requirements.in
django-filer==3.3.1
    # via -r requirements.in
django-filter==23.5
    # via wagtail
django-modelcluster==6.4
    # via wagtail
django-oauth-toolkit==3.0.1
    # via -r requirements.in
django-permissionedforms==0.1
    # via wagtail
django-polymorphic==3.1.0
    # via django-filer
django-select2-forms==3.0.0
    # via -r requirements.in
django-sortedm2m==4.0.0
    # via django-select2-forms
django-tables2==2.7.5
    # via -r requirements.in
django-taggit==4.0.0
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   wagtail
djangosaml2idp==0.7.2
    # via -r requirements.in
draftjs-exporter==2.1.7
    # via wagtail
easy-thumbnails==2.10
    # via django-filer
elementpath==4.8.0
    # via xmlschema
et-xmlfile==2.0.0
    # via openpyxl
filetype==1.2.0
    # via willow
fonttools==4.58.0
    # via matplotlib
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.2
    # via sqlalchemy
html2text==2025.4.15
    # via -r requirements.in
html5lib==1.1
    # via wagtail
icalendar==6.3.0
    # via
    #   caldav
    #   recurring-ical-events
    #   x-wr-timezone
idna==3.10
    # via
    #   requests
    #   yarl
importlib-resources==6.5.2
    # via matplotlib
iso3166==2.1.1
    # via qrbill
isort==6.0.1
    # via pylint
json-with-comments==1.2.10
    # via -r requirements.in
jwcrypto==1.5.6
    # via django-oauth-toolkit
kaleido==0.2.1
    # via -r requirements.in
kiwisolver==1.4.7
    # via matplotlib
kombu==5.5.3
    # via celery
l18n==2021.3
    # via wagtail
lxml==5.4.0
    # via
    #   caldav
    #   svglib
mailmanclient==3.3.5
    # via -r requirements.in
matplotlib==3.9.4
    # via -r requirements.in
mccabe==0.7.0
    # via pylint
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
mysqlclient==2.2.7
    # via -r requirements.in
narwhals==1.39.1
    # via plotly
numpy==2.0.2
    # via
    #   contourpy
    #   matplotlib
oauthlib==3.2.2
    # via
    #   django-oauth-toolkit
    #   requests-oauthlib
openpyxl==3.1.5
    # via
    #   -r requirements.in
    #   wagtail
packaging==25.0
    # via
    #   matplotlib
    #   plotly
    #   rocketchat-api
persistent==6.1.1
    # via
    #   btrees
    #   zodb
piecash==1.2.1
    # via -r requirements.in
pillow==10.4.0
    # via
    #   easy-thumbnails
    #   matplotlib
    #   pillow-heif
    #   reportlab
    #   wagtail
pillow-heif==0.22.0
    # via willow
platformdirs==4.3.8
    # via pylint
plotly==6.0.1
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via click-repl
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via redis
pylint==3.3.8
    # via -r requirements.in
pyopenssl==24.2.1
    # via pysaml2
pyparsing==3.2.3
    # via matplotlib
pypdf==5.5.0
    # via -r requirements.in
pysaml2==7.5.2
    # via djangosaml2idp
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   celery
    #   icalendar
    #   matplotlib
    #   pysaml2
    #   recurring-ical-events
    #   vobject
python-stdnum==2.0
    # via qrbill
pytz==2025.2
    # via
    #   datetime
    #   django-datetime-widget
    #   djangosaml2idp
    #   l18n
    #   piecash
    #   pysaml2
    #   vobject
qrbill==1.1.0
    # via -r requirements.in
qrcode==8.2
    # via qrbill
recurring-ical-events==3.7.0
    # via caldav
redis==5.3.0
    # via celery
reportlab==4.4.1
    # via
    #   easy-thumbnails
    #   svglib
requests==2.32.3
    # via
    #   caldav
    #   django-oauth-toolkit
    #   mailmanclient
    #   pysaml2
    #   requests-oauthlib
    #   rocketchat-api
    #   vdirsyncer
    #   wagtail
requests-oauthlib==2.0.0
    # via -r requirements.in
rocketchat-api==1.35.1
    # via -r requirements.in
ruff==0.12.10
    # via -r requirements.in
six==1.17.0
    # via
    #   html5lib
    #   l18n
    #   python-dateutil
    #   vobject
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==1.4.54
    # via
    #   piecash
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.2
    # via piecash
sqlparse==0.5.3
    # via django
svglib==1.5.1
    # via easy-thumbnails
svgwrite==1.4.3
    # via qrbill
telepath==0.3.1
    # via wagtail
tinycss2==1.4.0
    # via
    #   cssselect2
    #   svglib
tomlkit==0.13.3
    # via pylint
transaction==5.0
    # via zodb
types-python-dateutil==2.9.0.20250516
    # via arrow
typing-extensions==4.13.2
    # via jwcrypto
tzdata==2025.2
    # via
    #   icalendar
    #   kombu
    #   recurring-ical-events
    #   x-wr-timezone
tzlocal==5.3.1
    # via piecash
urllib3==2.4.0
    # via requests
vdirsyncer==0.19.3
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
vobject==0.9.9
    # via caldav
wagtail==5.2.8
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
willow==1.6.3
    # via wagtail
x-wr-timezone==2.0.1
    # via recurring-ical-events
xmlschema==2.5.1
    # via pysaml2
yarl==1.20.0
    # via aiohttp
zc-lockfile==3.0.post1
    # via zodb
zconfig==4.2
    # via zodb
zipp==3.21.0
    # via importlib-resources
zodb==6.0.1
    # via appy
zodbpickle==4.2
    # via zodb
zope-deferredimport==5.0
    # via persistent
zope-interface==7.2
    # via
    #   btrees
    #   datetime
    #   persistent
    #   transaction
    #   zodb
    #   zope-proxy
zope-proxy==6.1
    # via zope-deferredimport

# The following packages are considered to be unsafe in a requirements file:
# setuptools
sepa==0.5.3+mst1
